feat: Update material and prefab properties
The code changes involve updating the material and prefab properties for the Bullet object. The _Glossiness and _Metallic properties of the material have been set to 0, resulting in a less glossy and metallic appearance. In the Bullet prefab, the TrailRenderer's m_Curve values have been adjusted to create a different trail effect. These changes improve the visual presentation of the Bullet object in the game.